requirements.txt
flask
flask_sqlalchemy
flask_login
mysqlclient
matplotlib
python-dotenv
pymysql